Split up loss func headers and impls
