+++ flashlight/fl/nn/modules/BinaryCrossEntropy.cpp
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+#include "flashlight/fl/nn/modules/BinaryCrossEntropy.h"
+
+#include <stdexcept>
+
+#include "flashlight/fl/autograd/Functions.h"
+
+namespace fl {
+
+Variable BinaryCrossEntropy::forward(
+    const Variable& inputs,
+    const Variable& targets) {
+  return mean(flat(binaryCrossEntropy(inputs, targets)), {0});
+}
+
+Variable BinaryCrossEntropy::forward(
+    const Variable& inputs,
+    const Variable& targets,
+    const Variable& weights) {
+  return mean(flat(weights * binaryCrossEntropy(inputs, targets)), {0});
+}
+
+std::string BinaryCrossEntropy::prettyString() const {
+  return "BinaryCrossEntropy";
+}
+
+} // namespace fl
